Use resolution for glossy in package.json with tz fix
There is a bug in glossy (syslog formatter) that causes the date to be
generated in UTC, but with the offset of the local timezone included.
This affects the recorded logs when using the RFC5424 syslog type.
Winston-syslog depends on glossy and there is a partial fix in the
upstream [squeeks/glossy](https://github.com/squeeks/glossy) repo.
However, the year is still wrong and the fix on github has not been
released to npm since 2 years. Manually resolving the dependency to a
fixed git commit seems like a decent workaround until upstream is
fixed and released.