DX: Set node.js version range + validate it
Check that no dependency has a more narrow version range than we do. It ignores eslint as eslint has a way too narrow node.js range to be able to test against like this. It would be too verbose.

Defaults to "8.0.0" as the .travis.yml only references the latest LTS, which is 10