Handle correlation IDs automatically
This is an attempt at automatically handling logging of correlation IDs
and other debug metadata. This is implemented with inspiration from
https://dev.to/elmatella/my-logging-strategy-for-express-1mk8. We store
the metadata in an `AsyncLocalStorage`, which is local to each
`async` context, and provide a middleware factory for Express which sets
this from the request (which will be used in `lu-server`) and then
logging it out automatically (currently under a temporary label
`correlationIdTest`).